requirements.txt
flask==2.3.3
python-dotenv==1.0.0
requests==2.31.0
Werkzeug==2.3.7
Flask-Session==0.5.0
psycopg2-binary==2.9.9
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.23
Flask-Migrate==4.0.5 